refactor: modularize dashboard components
